+++ src/object/SuperObject.java
@@ -18,8 +18,8 @@ public class SuperObject {
 	protected UtilityTool uTool = new UtilityTool();
 	
 	public void draw(Graphics2D g2, GamePanel gp) {
-		int screenX = worldX - gp.player.worldX + gp.player.screenX;
-		int screenY = worldY - gp.player.worldY + gp.player.screenY;
+		int screenX = (int) worldX - (int) gp.player.worldX + gp.player.screenX;
+		int screenY = (int) worldY - (int) gp.player.worldY + gp.player.screenY;
 
 		if (worldX + gp.tileSize > gp.player.worldX - gp.player.screenX 
 			&& worldX - gp.tileSize < gp.player.worldX + gp.player.screenX
